Execute rac_command off of main thread

+++ ExampleProject/PrimesViewController.m
@@ -58,9 +58,10 @@
         NSInteger latestPrime = [x integerValue];
         if (latestPrime == -1) {
             self.primesLog.text = @"";
+
             return;
         }
-        NSString *latest = [NSString stringWithFormat:@"%d, ", latestPrime];
+        NSString *latest = [NSString stringWithFormat:@"%ld, ", (long)latestPrime];
         self.primesLog.text = [self.primesLog.text stringByAppendingString:latest];
     }];
 }
